+++ src/components/Summary/index.tsx
@@ -1,3 +1,6 @@
+import { useContext } from 'react';
+
+import { TransactionsContext } from '../../TransactionsContext';
 import { Container } from './styles';
 
 import incomeImg from '../../assets/income.svg';
@@ -5,6 +8,8 @@ import outcomeImg from '../../assets/outcome.svg';
 import totalImg from '../../assets/total.svg';
 
 export function Summary() {
+  const transactions = useContext(TransactionsContext);
+
   return (
     <Container>
       <div>
